Move riiif dependent code to the generated riiif initializer
